:hammer: Moved the Guided Tour in the Student Home Page to its own component
