+++ app/imports/ui/components/student/HomePage/StudentHomeRecommendedTeasers.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Header } from 'semantic-ui-react';
+
+interface IStudentHomeRecommendedTeasersProps {
+
+}
+
+const StudentHomeRecommendedTeasers = (props: IStudentHomeRecommendedTeasersProps) => {
+  return (
+    <>
+      <Header dividing>RECOMMENDED</Header>
+    </>
+  );
+};
+
+export default StudentHomeRecommendedTeasers;
